Fixes SDM loopy and moves Differentiable out of transforms
