+++ menpofit/transform/__init__.py
@@ -1,7 +1,7 @@
 from .modeldriven import ModelDrivenTransform, OrthoMDTransform
-from .differentiable import DL, DP, DX
 from .homogeneous import (DifferentiableAffine, DifferentiableSimilarity,
-                          DifferentiableAlignmentSimilarity)
+                          DifferentiableAlignmentSimilarity,
+                          DifferentiableAlignmentAffine)
 from .piecewiseaffine import DifferentiablePiecewiseAffine
 from .thinsplatesplines import DifferentiableThinPlateSplines
 from .rbf import DifferentiableR2LogR2RBF, DifferentiableR2LogRRBF
